Bug fix: invoke gphoto with only one --port option
Current code modifies the (global) gphoto command line array,
inserting "--port ptpip:NNNN". This is happening on each
iteration, so not only is the command line growing infinitely
but it's also growing in the wrong direction: the port arg is
inserted at the beginning, hence is overridden by the older
but later-in-the-command-line options.

Solution: make a new copy of the command + args each time we invoke.

Signed-off-by: Ed Santiago <ed@edsantiago.com>